+++ fabric/src/main/java/dev/tapetum/shaders/mixin/AccessorGlBuffer.java
@@ -1,0 +1,10 @@
+package dev.tapetum.shaders.mixin;
+
+import com.mojang.blaze3d.opengl.GlBuffer;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+@Mixin(GlBuffer.class)
+public interface AccessorGlBuffer {
+    @Accessor("handle") int tapetum$handle();
+}